Added Gradle to project, updated filepaths to images
