+++ PawfectPairs/build.gradle.kts
@@ -1,0 +1,49 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    `java-library`
+    id("application")
+    id("org.openjfx.javafxplugin") version "0.1.0"
+    
+}
+
+javafx {
+    version = "21"
+    modules("javafx.controls", "javafx.fxml")
+}
+
+repositories {
+    mavenLocal()
+    maven {
+        url = uri("https://repo.maven.apache.org/maven2/")
+    }
+}
+
+dependencies {
+    
+}
+
+sourceSets {
+    main {
+        java {
+            setSrcDirs(listOf("src"))
+        }
+    }
+
+    test {
+        java {
+            setSrcDirs(listOf("test"))
+        }
+    }
+}
+
+application {
+  mainClass.set("guilayout.ApplicationStart")  
+}
+
+group = "PawfectPairs"
+version = "0.0.1-SNAPSHOT"
+description = "PawfectPairs"
+java.sourceCompatibility = JavaVersion.VERSION_1_8
