+++ PawfectPairs/gradle/libs.versions.toml
@@ -1,0 +1,2 @@
+# This file was generated by the Gradle 'init' task.
+# https://docs.gradle.org/current/userguide/platforms.html#sub::toml-dependencies-format
