+++ PawfectPairs/src/guicontrol/DogProfileController.java
@@ -6,22 +6,19 @@ import java.util.List;
 import backend.dog.Dog;
 import backend.poster.Poster;
 import guilayout.Components;
-import javafx.scene.control.Hyperlink;
-import javafx.scene.control.Label;
+import javafx.scene.control.*;
 
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.text.Text;
-import javafx.scene.text.TextFlow;
-import javafx.stage.Stage;
+import javafx.scene.image.*;
+import javafx.scene.text.*;
+import javafx.stage.*;
 
 public class DogProfileController {
     private int currentIndex = 0;
     private ArrayList<Dog> dogProfiles = new ArrayList<>(List.of(
     
-    		new Dog("Charles", 1, 2, 1, 2, "M", new Poster(5, "doglover123", 12), false, "/guilayout/dog.png", "Charles is involed in some shady stuff. Don't ask questions you don't want to know the answers to."), 
-    		new Dog("Sandy", 2, 3, 1, 2, "F", new Poster(5, "doglover123", 12), false,"/guilayout/dog1.png", "Sandy looks like she knows something you don't... but that's what makes her so loveable!"),
-    		new Dog("Chuckles", 3, 8, 0, 0, "M", new Poster(5, "doglover123", 12), false,"/guilayout/dog4.png", "Lazy and old Chuckles is the perfect lap dog for anyone who doesn't like to move.")
+    		new Dog("Charles", 1, 2, 1, 2, "M", new Poster(5, "doglover123", 12), false, "file:src/guilayout/dog.png", "Charles is involed in some shady stuff. Don't ask questions you don't want to know the answers to."), 
+    		new Dog("Sandy", 2, 3, 1, 2, "F", new Poster(5, "doglover123", 12), false,"file:src/guilayout/dog1.png", "Sandy looks like she knows something you don't... but that's what makes her so loveable!"),
+    		new Dog("Chuckles", 3, 8, 0, 0, "M", new Poster(5, "doglover123", 12), false,"file:src/guilayout/dog4.png", "Lazy and old Chuckles is the perfect lap dog for anyone who doesn't like to move.")
 
 ));
